Update dialog creation in  ShowAskBoxToSetOrUnset().
Updates the creation of 'SetUnsetDialog', also removes deprecation
warnings.
We are now closing the dialog in the same functions it's
being created in. This means that we can move the variable
'SetUnsetDialog' from global to local scope.

Changed the function parameters for  DoFunctionSetOrUnsetVar(),
we can now send in an int instead of a void pointer.
Removed unnecessary cast.

Some comments was also removed.

There should be no functional change in this commit.